# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: This is a training server
  version: 1.0.0
  title: NodeJS
host: localhost:8080
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /swagger:
    x-swagger-pipe: swagger_raw
  /cities:
    x-swagger-router-controller: cities
    get:
      operationId: getAllCities
      tags:
      - Cities
      summary: Get all cities
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/City'
        404:
          description: Error
    post:
      operationId: addNewCity
      tags:
        - Cities
      summary: Adds new city and returns it
      parameters:
        - in: body
          name: city
          required: true
          description: The city to add
          schema:
            type: object
            required:
              - name
              - country
              - capital
            properties:
              name:
                type: string
                example: 'New York'
              country:
                type: string
                example: 'USA'
              capital:
                type: boolean
                example: false
              location:
                type: object
                properties:
                  lat:
                    type: number
                    example: 56.86186
                  long:
                    type: number
                    example: 53.232428
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/City'
        404:
          description: Error
  /cities/{id}:
    x-swagger-router-controller: cities
    put:
      operationId: upsertCity
      tags:
        - Cities
      summary: Updates existing city or adds new one with given ID
      parameters:
        - name: id
          in: path
          description: ID of the city to upsert
          required: true
          type: string
        - in: body
          name: city
          required: true
          description: The city to upsert
          schema:
            type: object
            required:
              - name
              - country
              - capital
            properties:
              name:
                type: string
                example: 'New York'
              country:
                type: string
                example: 'USA'
              capital:
                type: boolean
                example: false
              location:
                type: object
                properties:
                  lat:
                    type: number
                    example: 56.86186
                  long:
                    type: number
                    example: 53.232428
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/City'
        404:
          description: Error
    delete:
      operationId: deleteCity
      tags:
        - Cities
      summary: Deletes city with given ID
      parameters:
        - name: id
          in: path
          description: ID of the city to delete
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: Error
  /cities/random:
    x-swagger-router-controller: cities
    get:
      operationId: getRandomCity
      tags:
      - Cities
      summary: Get random city
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/City'
        404:
          description: Error
  /users:
    x-swagger-router-controller: users
    get:
      operationId: getAllUsers
      tags:
      - Users
      summary: Get all users
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        404:
          description: Error
  /users/{id}:
    x-swagger-router-controller: users
    put:
      operationId: upsertUser
      tags:
        - Users
      summary: Updates existing user or adds new one with given ID
      parameters:
        - name: id
          in: path
          description: ID of the user to upsert
          required: true
          type: string
        - in: body
          name: user
          required: true
          description: The user to upsert
          schema:
            type: object
            required:
              - firstName
              - lastName
              - gender
            properties:
              firstName:
                type: string
                example: 'Ross'
              lastName:
                type: string
                example: 'Geller'
              gender:
                type: boolean
                example: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
        404:
          description: Error
    delete:
      operationId: deleteUser
      tags:
        - Users
      summary: Deletes user with given ID
      parameters:
        - name: id
          in: path
          description: ID of the user to delete
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: Error
  /products:
    x-swagger-router-controller: products
    get:
      operationId: getAllProducts
      tags:
      - Products
      summary: Get all products
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
        404:
          description: Error
    post:
      operationId: addNewProduct
      tags:
        - Products
      summary: Adds new product and returns it
      parameters:
        - in: body
          name: product
          required: true
          description: The product to add
          schema:
            type: object
            required:
              - manufacturer
              - model
            properties:
              manufacturer:
                type: string
                example: 'Apple'
              model:
                type: string
                example: 'iPhone'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product'
        404:
          description: Error
  /products/{id}:
    x-swagger-router-controller: products
    get:
      operationId: getProduct
      tags:
        - Products
      summary: Get product with given ID
      parameters:
        - name: id
          in: path
          description: ID of the product to get
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product'
        404:
          description: Error
    put:
      operationId: upsertProduct
      tags:
        - Products
      summary: Updates existing product or adds new one with given ID
      parameters:
        - name: id
          in: path
          description: ID of the product to upsert
          required: true
          type: string
        - in: body
          name: product
          required: true
          description: The product to upsert
          schema:
            type: object
            required:
              - manufacturer
              - model
            properties:
              manufacturer:
                type: string
                example: 'Apple'
              model:
                type: string
                example: 'iPhone'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product'
        404:
          description: Error
    delete:
      operationId: deleteProduct
      tags:
        - Products
      summary: Deletes product with given ID
      parameters:
        - name: id
          in: path
          description: ID of the product to delete
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: Error
  /products/{id}/reviews:
    x-swagger-router-controller: products
    get:
      operationId: getProductReviews
      tags:
      - Products
      summary: Get all reviews for the product with given ID
      parameters:
        - name: id
          in: path
          description: ID of the product to get the reviews for
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              _id:
                type: string
                example: '123456789012345678901234'
              reviews:
                type: array
                items:
                  type: string
                  example: 'iPhone rocks!'
        404:
          description: Error
definitions:
  City:
    type: object
    properties:
      _id:
        type: string
        example: '123456789012345678901234'
      name:
        type: string
        example: 'New York'
      country:
        type: string
        example: 'USA'
      capital:
        type: boolean
        example: false
      location:
        type: object
        properties:
          lat:
            type: number
            minimum: -90
            maximum: 90
            example: 56.86186
          long:
            type: number
            minimum: -180
            maximum: 180
            example: 53.232428
  User:
    type: object
    properties:
      _id:
        type: string
        example: '123456789012345678901234'
      firstName:
        type: string
        example: 'Ross'
      lastName:
        type: string
        example: 'Geller'
      gender:
        type: boolean
        example: true
  Product:
    type: object
    properties:
      _id:
        type: string
        example: '123456789012345678901234'
      manufacturer:
        type: string
        example: 'Apple'
      model:
        type: string
        example: 'iPhone'